# docs/openapi.yaml
openapi: 3.0.3
info:
  title: OCMS & Store LaPet Omnichannel Master API
  description: >
    Hợp đồng giao tiếp API chuẩn hóa toàn diện (Master OpenAPI Specification) giữa
    Hệ thống Quản trị Vận hành Đa kênh OCMS và Website Bán Hàng / Cổng Đại lý (store-lapet),
    đồng thời tích hợp liên thông với ERP Odoo và Kênh Zalo OA.
    
    Tài liệu này bao quát toàn bộ các nhóm tính năng vận hành:
    1. **Orders (Đơn hàng & Tiếp nhận chuẩn):** Cổng tiếp nhận đơn hàng bất đồng bộ, tra cứu trạng thái, duyệt đơn Odoo, hủy đơn và thống kê.
    2. **Products & Inventory (Sản phẩm & Tồn kho):** Danh mục sản phẩm, bảng giá đại lý/bán lẻ, kiểm tra tồn kho Odoo thời gian thực.
    3. **Promotions & Pricing Engine (Khuyến mãi & Động cơ tính giá):** Động cơ tính chiết khấu bậc thang tự động cho giỏ hàng, tư vấn ưu đãi bằng AI.
    4. **Customers & Dealers (Khách hàng & Đại lý B2B):** Xác thực OTP/Mật khẩu, hồ sơ đại lý, hạn mức công nợ, doanh số và quản lý danh bạ CRM.
    5. **Omnichannel Chat (Hội thoại & Zalo):** Quản lý luồng trò chuyện, gửi tin nhắn trao đổi, tự động bóc tách đơn hàng AI từ chat.
    6. **Webhooks & Sync (Thanh toán & Đồng bộ):** Webhook thanh toán tự động VietQR và kích hoạt đồng bộ dữ liệu Odoo ERP.
  version: 2.0.0
  contact:
    name: OCMS Integration Architecture Team
    email: dev@lapet.vn
servers:
  - url: http://localhost:3000
    description: Môi trường phát triển cục bộ (Local Development)
  - url: https://api.ocms.lapet.vn
...
paths:
...
  /api/v1/conversations/{id}/messages:
    get:
      tags:
        - Chat & Zalo
      summary: Lịch sử tin nhắn trong một cuộc trò chuyện
      operationId: getConversationMessages
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Danh sách tin nhắn

    post:
      tags:
        - Chat & Zalo
      summary: Gửi tin nhắn phản hồi cho khách hàng qua Zalo
      operationId: sendChatMessage
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content
              properties:
                content:
                  type: string
                  example: "Dạ em đã nhận được yêu cầu của anh chị ạ!"
      responses:
        '200':
          description: Tin nhắn đã được gửi thành công

  # ============================================================================
  # 6. WEBHOOKS & SYNC (THANH TOÁN & ĐỒNG BỘ)
  # ============================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...